Start working on consolidating completion and chat_completion underlying implementations
