+++ mistralrs-server-core/src/util.rs
@@ -1,11 +1,62 @@
 //! ## General utilities.
 
+use anyhow::{Context, Result};
+use axum::{extract::Json, http::StatusCode, response::IntoResponse};
 use image::DynamicImage;
-use mistralrs_core::AudioInput;
+use mistralrs_core::Request;
+use serde::Serialize;
 use tokio::{
     fs::{self, File},
     io::AsyncReadExt,
 };
+
+use crate::types::SharedMistralRsState;
+
+/// Trait for converting errors to HTTP responses with appropriate status codes.
+pub(crate) trait ErrorToResponse: Serialize {
+    /// Converts the error to an HTTP response with the specified status code.
+    fn to_response(&self, code: StatusCode) -> axum::response::Response {
+        let mut r = Json(self).into_response();
+        *r.status_mut() = code;
+        r
+    }
+}
+
+/// Standard JSON error response structure.
+#[derive(Serialize)]
+pub(crate) struct JsonError {
+    pub(crate) message: String,
+}
+
+impl JsonError {
+    /// Creates a new JSON error with the specified message.
+    pub(crate) fn new(message: String) -> Self {
+        Self { message }
+    }
+}
+impl ErrorToResponse for JsonError {}
+
+/// Internal error type for model-related errors with a descriptive message.
+///
+/// This struct wraps error messages from the underlying model and implements
+/// the standard error traits for proper error handling and display.
+#[derive(Debug)]
+pub(crate) struct ModelErrorMessage(pub(crate) String);
+impl std::fmt::Display for ModelErrorMessage {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+impl std::error::Error for ModelErrorMessage {}
+
+/// Sends a request to the model processing pipeline.
+pub async fn send_model_request(state: &SharedMistralRsState, request: Request) -> Result<()> {
+    let sender = state
+        .get_sender()
+        .context("mistral.rs sender not available.")?;
+
+    sender.send(request).await.map_err(|e| e.into())
+}
 
 /// Parses and loads an image from a URL, file path, or data URL.
 ///
@@ -81,46 +132,6 @@ pub async fn parse_image_url(url_unparsed: &str) -> Result<DynamicImage, anyhow:
     Ok(image::load_from_memory(&bytes)?)
 }
 
-/// Parses and loads an audio file from a URL, file path, or data URL.
-pub async fn parse_audio_url(url_unparsed: &str) -> Result<AudioInput, anyhow::Error> {
-    let url = if let Ok(url) = url::Url::parse(url_unparsed) {
-        url
-    } else if File::open(url_unparsed).await.is_ok() {
-        url::Url::from_file_path(std::path::absolute(url_unparsed)?)
-            .map_err(|_| anyhow::anyhow!("Could not parse file path: {}", url_unparsed))?
-    } else {
-        url::Url::parse(url_unparsed)
-            .map_err(|_| anyhow::anyhow!("Could not parse as base64 data: {}", url_unparsed))?
-    };
-
-    let bytes = if url.scheme() == "http" || url.scheme() == "https" {
-        match reqwest::get(url.clone()).await {
-            Ok(http_resp) => http_resp.bytes().await?.to_vec(),
-            Err(e) => anyhow::bail!(e),
-        }
-    } else if url.scheme() == "file" {
-        let path = url
-            .to_file_path()
-            .map_err(|_| anyhow::anyhow!("Could not parse file path: {}", url))?;
-
-        if let Ok(mut f) = File::open(&path).await {
-            let metadata = fs::metadata(&path).await?;
-            let mut buffer = vec![0; metadata.len() as usize];
-            f.read_exact(&mut buffer).await?;
-            buffer
-        } else {
-            anyhow::bail!("Could not open file at path: {}", url);
-        }
-    } else if url.scheme() == "data" {
-        let data_url = data_url::DataUrl::process(url.as_str())?;
-        data_url.decode_to_vec()?.0
-    } else {
-        anyhow::bail!("Unsupported URL scheme: {}", url.scheme());
-    };
-
-    AudioInput::from_bytes(&bytes)
-}
-
 #[cfg(test)]
 mod tests {
     use image::GenericImageView;
@@ -191,12 +202,5 @@ mod tests {
         let url = format!("data:image/png;base64,{}", url);
         let image = parse_image_url(&url).await.unwrap();
         assert_eq!(image.dimensions(), (32, 32));
-
-        // audio from base64
-        let audio_b64 = "UklGRiYAAABXQVZFZm10IBAAAAABAAEAQB8AAIA+AAACABAAZGF0YQIAAAAAAA==";
-        let url = format!("data:audio/wav;base64,{audio_b64}");
-        let audio = parse_audio_url(&url).await.unwrap();
-        assert_eq!(audio.sample_rate, 8000);
-        assert_eq!(audio.samples.len(), 1);
     }
 }
